optimize info pages
